% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_boxplot.R
\name{disease_boxplot}
\alias{disease_boxplot}
\title{Produce Boxplots of Disease Survey Data for a Given Crop}
\usage{
disease_boxplot(fd, crop_graph)
}
\arguments{
\item{crop_graph}{Crop of interest to produce a graph for.

#' @section Accepted \code{crop} Values:
 Acceptable values for crop are
 \itemize{
   \item \code{barley},
   \item \code{canola},
   \item \code{chickpea},
   \item \code{maize},
   \item \code{mungbean},
   \item \code{peanut},
   \item \code{sorghum},
   \item \code{soybean},
   \item \code{sunflower},
   \item \code{wheat}
 }}

\item{survey}{Fulcrum data retrieved using \code{get_fulcrum}.}
}
\description{
Produce Boxplots of Disease Survey Data for a Given Crop
}
\author{
adam.sparks@usq.edu.au
}
