% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fulcrum.R
\name{fetch_fulcrum}
\alias{fetch_fulcrum}
\title{Fetches Data from Fulcrum and Formats it for Use in Reporting}
\usage{
fetch_fulcrum()
}
\description{
Fetches Data from Fulcrum and Formats it for Use in Reporting
}
