% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fulcrum.R
\name{filter_fulcrum}
\alias{filter_fulcrum}
\title{Filters Survey Data from get_fulcrum}
\usage{
filter_fulcrum(fd, crop = NULL, disease = NULL,
  location_description = NULL, season = NULL)
}
\arguments{
\item{fd}{Required. Fulcrum data to be filtered.  Data must be a result of
\code{\link{get_fulcrum}}.}

\item{crop}{Optional. Character vector of crop(s) to filter on.}

\item{disease}{Optional. Character vector of disease(s) to filter on.}

\item{season}{Optional. Character vector of season(s) to filter on.}

\item{location}{Optional. Character vector of location(s) to filter on.}
}
\value{
A \code{\link[tibble]{tibble}} of filtered Fulcrum data by the
 requested parameters.
}
\description{
Filters data retrieved using \code{\link{get_fulcrum}}.  Data can be filtered
by locations surveyed, crop, disease, location description or season.
}
\section{Accepted \code{crop} Values}{

 Acceptable values for crop are
 \itemize{
   \item \code{barley},
   \item \code{canola},
   \item \code{chickpea},
   \item \code{maize},
   \item \code{mungbean},
   \item \code{peanut},
   \item \code{sorghum},
   \item \code{soybean},
   \item \code{sunflower},
   \item \code{wheat}
 }
}

\section{Accepted \code{disease} Values}{

 Acceptable values for disease are
 \itemize{
   \item \code{},
 }
}

\section{Accepted \code{location_description} Values}{

 Acceptable values for location are
 \itemize{
   \item \code{farm field},
   \item \code{commercial trial},
   \item \code{research trial} or
   \item \code{roadside}.
 }
}

\section{Accepted \code{season} Values}{

 Acceptable values for season are
 \itemize{
   \item \code{Winter 2018},
   \item \code{Summer 2018/19},
   \item \code{Winter 2019} or
   \item \code{Summer 2019/20}.
 }
}

\examples{
\dontrun{
# map only mungbean surveys
x <- filter_fulcrum(fd = fulcrum_data, crop = "mungbean")
}
}
\seealso{
\code{\link{get_fulcrum}} for retrieving and formatting Fulcrum
 data
}
\author{
Sparks, Adam H. \email{adam.sparks@usq.edu.au}
}
