% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Load your Fulcrum API key from file}
\usage{
get_api_key(key_location)
}
\arguments{
\item{key_location}{File location of your API key in .RData format}
}
\value{
API key in plain text to be passed to the API query functions
}
\description{
This function is to help obscure the API key, and prevent it showing up in
logs or accidentally being distributed. The function expects the file to
already be in .RData format. Do not assign the output of this function to a
variable because that will defeat the point of putting the key into a
function in the first place.
}
