% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fulcrum.R
\name{get_fulcrum}
\alias{get_fulcrum}
\title{Fetches Data from Fulcrum.app and Formats it for Use in Reporting}
\usage{
get_fulcrum()
}
\value{
A tidy data frame \code{tibble} object of Fulcrum survey data
}
\description{
Fetch and clean Fulcrum data from USQ/DAFQ in-paddock surveys of crop
disease incidence and oher related notes for reporting.
}
\examples{
\donttest{
fd <- get_fulcrum()
}
}
