% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Fetches Data from Fulcrum.app and Returns Only Locations and Location Information}
\usage{
get_locations(fulcrum_url = NULL)
}
\arguments{
\item{fulcrum_url}{A url provided by \url{https://www.fulcrumapp.com/} for sharing a
'CSV' format fill of Fulcrum data}
}
\value{
A \code{\link[sf]{sf}} object of locations and location information,
but no disease data.
}
\description{
Fetch and clean Fulcrum data from USQ/DAFQ in-paddock surveys and create
a dataframe of locations and crops surveyed with no disease data
}
\examples{

\dontrun{
# using a URL set in the .Renviron, get data from Fulcrum
x <- get_locations()
}

\dontrun{
# specify a url to get data from Fulcrum
x <- get_locations(fulcrum_url = "https://web.fulcrumapp.com/shares/########.csv")
}

}
\author{
Sparks, Adam H. \email{adam.sparks@usq.edu.au}
}
