% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_tibble2.R
\name{json_to_tibble2}
\alias{json_to_tibble2}
\title{A better version of \code{json_to_tibble} to convert raw JSON format file to an R-friendly tibble}
\usage{
json_to_tibble2(json_obj)
}
\arguments{
\item{json_obj}{A JSON formatted object, generated by \code{query_fulcrum}}
}
\value{
The datatable section of the JSON file as a tibble.
}
\description{
\code{json_to_tibble2} creates an R-friendly tibble() object from the raw JSON
data extracted by \code{query_fulcrum}. It discards the metadata included
in the JSON payload which is probably not useful information for our
analysis. Use the output dataframes for subsequent analysis and checking.
Don't overwrite the objects you create with this function during data
wrangling as they should be used as the base dfs to run all checks and build
the final data outputs in RShiny.
}
