% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fulcrum.R
\name{map_fulcrum}
\alias{map_fulcrum}
\title{Maps Locations Surveyed Using Data from get_fulcrum}
\usage{
map_fulcrum(fd, crop = NULL, disease = NULL, location = NULL,
  season = NULL)
}
\arguments{
\item{fd}{Required. Fulcrum data to use for mapping. Must be supplied from
\code{\link{get_fulcrum}}.}

\item{crop}{Optional. Map by crop type. If not specified all crops are
mapped.}

\item{disease}{Optional. Map by disease of interest. If not specified, all
diseases are mapped. Overrides crop type.}

\item{location}{Optional. Map by location type. If not specified all
locations are mapped.}

\item{season}{Optional. Map by season. If not specified, all seasons mapped.}
}
\value{
A map of Fulcrum data
}
\description{
Generates maps of locations surveyed using \code{\link{get_fulcrum}} data.
Maps can be of all locations surveyed, or by crop, disease, location
 description or season.
}
\section{Crop Values}{

 Acceptable values for crop are
 \itemize{
   \item \code{barley},
   \item \code{canola},
   \item \code{chickpea},
   \item \code{maize},
   \item \code{mungbean},
   \item \code{peanut},
   \item \code{sorghum},
   \item \code{soybean},
   \item \code{sunflower},
   \item \code{wheat}
 }
}

\section{Disease Values}{

 Acceptable values for disease are
 \itemize{
   \item \code{},
 }
}

\section{Location Values}{

 Acceptable values for location are
 \itemize{
   \item \code{farm field},
   \item \code{commercial trial},
   \item \code{research trial} or
   \item \code{roadside}.
 }
}

\section{Season Values}{

 Acceptable values for season are
 \itemize{
   \item \code{Winter 2018},
   \item \code{Summer 2018/19},
   \item \code{Winter 2019} or
   \item \code{Summer 2019/20}.
 }
}

\examples{

\dontrun{
# map all surveys in Fulcrum database
fulcrum_data <- get_fulcrum()
map <- map_fulcrum(fd = fulcrum_data)
}

\dontrun{
# map only mungbean surveys
map <- map_fulcrum(fd = fulcrum_data, crop = "mungbean")
}
}
\seealso{
\code{\link{get_fulcrum}} for retrieving and formatting Fulcrum
 data
}
\author{
Sparks, Adam H. \email{adam.sparks@usq.edu.au}
}
