% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_outline.R
\docType{data}
\name{oz_outline}
\alias{oz_outline}
\title{Outline of Queensland and New South Wales}
\format{A \code{\link[sf]{sf}} object.}
\usage{
oz_outline
}
\description{
Data to provide a map for reporting, this data provides a layer with
a federation-level outline for use on top of \code{\link{oz_label}}.
}
\seealso{
\code{\link{oz_label}}
}
\keyword{datasets}
