% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_sf.R
\docType{data}
\name{oz_sf}
\alias{oz_sf}
\title{Base Map of Queensland and New South Wales}
\format{A \code{\link[sf]{sf}} object.}
\usage{
oz_sf
}
\description{
Data to provide a basemap for reporting.
}
\keyword{datasets}
