% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_fulcrum.R
\name{query_fulcrum}
\alias{query_fulcrum}
\title{Pass a GET query to the Fulcrum Query API}
\usage{
query_fulcrum(
  query,
  api_key,
  url = "https://api.fulcrumapp.com/api/v2/query/",
  format = "json"
)
}
\arguments{
\item{query}{The path to a PostgreSQL query saved in a text file, or the query itself as a string}

\item{api_key}{API key for your Fulcrum account in .RData format.}

\item{url}{URL for the query API. Default should be fine.}

\item{format}{Output format for the query.}
}
\value{
The output of the SQL query from Fulcrum in JSON format.
}
\description{
Fulcrum uses PostgreSQL on the backend, and so any valid PostgreSQL query can
be passed to the API. Save an SQL query in a text file and pass that query to
\code{query_fulcrum_api} and, assuming you have a valid API key, the results
of the query will be returned in JSON format for further processing in R.
}
\details{
This function depends on the \code{httr} library to work.

\code{query_fulcrum_api} expects your API key to be input as a .RData file.
To convert a plain text string to a .RData file, use
\code{\link[readRDS:base]{saveRDS}} from base R.
}
