% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fulcrum.R
\name{filter_fulcrum}
\alias{filter_fulcrum}
\title{Filters Survey Data from get_fulcrum}
\usage{
filter_fulcrum(fd, crop = NULL, disease = NULL,
  location_description = NULL, season = NULL)
}
\arguments{
\item{fd}{Required. Fulcrum data to be filtered.  Data must be a result of
\code{\link{get_fulcrum}}.}

\item{crop}{Optional. Character vector of crop(s) to filter on.}

\item{disease}{Optional. Character vector of disease(s) to filter on.}

\item{location_description}{Optional. Character vector of location(s) to filter on.}

\item{season}{Optional. Character vector of season(s) to filter on.}
}
\value{
A \code{\link[tibble]{tibble}} of filtered Fulcrum data by the
 requested parameters.
}
\description{
Filters data retrieved using \code{\link{get_fulcrum}}.  Data can be filtered
by locations surveyed, crop, disease, location description or season.
}
\section{Accepted \code{crop} Values}{

 Acceptable values for crop are
 \itemize{
   \item \code{barley},
   \item \code{canola},
   \item \code{chickpea},
   \item \code{maize},
   \item \code{mungbean},
   \item \code{peanut},
   \item \code{sorghum},
   \item \code{soybean},
   \item \code{sunflower},
   \item \code{wheat}
 }
}

\section{Accepted \code{disease} Values}{

 Acceptable values for disease are
 \itemize{
   \item \code{Fusarium head blightscab},
   \item \code{Net form net blotch},
   \item \code{Sspot form net blotch},
   \item \code{Powdery mildew},
   \item \code{Dtem rust},
   \item \code{Virus},
   \item \code{Powdery mildew},
   \item \code{Blackleg},
   \item \code{Sclerotinia stem rot},
   \item \code{Virus},
   \item \code{Ascochyta blight},
   \item \code{Botrytis grey mould},
   \item \code{Fusarium wilt},
   \item \code{Phytoplasma},
   \item \code{Phytophthora root rot},
   \item \code{Bacterial stalk rot and bacterial top rot},
   \item \code{Charcoal rot},
   \item \code{Common rust},
   \item \code{Commonboil smut},
   \item \code{Crazy top},
   \item \code{Damping off},
   \item \code{Diplodia cob rot},
   \item \code{Downy mildew},
   \item \code{Dwarf mosaic virus},
   \item \code{Fusarium ear rot},
   \item \code{Fusarium stalk rot},
   \item \code{Head smut},
   \item \code{Pythium stalk rot},
   \item \code{Turcica leaf blight},
   \item \code{Maydis leaf blight},
   \item \code{Southern rust},
   \item \code{Alternaria leaf spot},
   \item \code{Bacterial blight},
   \item \code{Charcoal rot},
   \item \code{Damping off},
   \item \code{Fusarium wilt},
   \item \code{Gummy pod},
   \item \code{Halo blight},
   \item \code{Phytoplasma},
   \item \code{Powdery mildew},
   \item \code{Puffy pod},
   \item \code{Rhizoctonia rot},
   \item \code{Root lesion nematode},
   \item \code{Sclerotinia stem rot},
   \item \code{Sclerotium stem rot},
   \item \code{Tan spot},
   \item \code{Tsv},
   \item \code{Fusarium root rot},
   \item \code{Neocosmospora root rot},
   \item \code{Net blotch},
   \item \code{Kernel shrivel syndrome},
   \item \code{Root lesion nematode},
   \item \code{Rust},
   \item \code{Sclerotium base rot},
   \item \code{Bacterial top and stalk rot},
   \item \code{Damping off},
   \item \code{Ergot},
   \item \code{Fusarium head blight},
   \item \code{Fusarium stalk rot},
   \item \code{Grain mould},
   \item \code{Head smut},
   \item \code{Johnsongrass mosaic virus},
   \item \code{Leaf blight},
   \item \code{Root lesion nematode},
   \item \code{Sclerotium base rot},
   \item \code{Tar spot},
   \item \code{Bacterial blight bacterial pustule},
   \item \code{Damping off},
   \item \code{Peanut mottle virus},
   \item \code{Phomopsis seed decay},
   \item \code{Phytophthora root stem and root rot},
   \item \code{Phytoplasma},
   \item \code{Pod stem cankerblight},
   \item \code{Purple seed stain},
   \item \code{Rhizoctonia rot},
   \item \code{Root lesion nematode},
   \item \code{Sclerotinia rot},
   \item \code{Soybean mosaic virus},
   \item \code{Apical chlorosis},
   \item \code{Botrytis head rot grey mould},
   \item \code{Rhizopus head rot},
   \item \code{Sclerotinia rot},
   \item \code{Sclerotium base rot},
   \item \code{Stem cankerblight},
   \item \code{Verticillium wilt},
   \item \code{Crown rot},
   \item \code{Common root rot},
   \item \code{Leafbrown rust},
   \item \code{Root lesion nematode},
   \item \code{Septoria nodorum blotch},
   \item \code{Stemblack rust},
   \item \code{Stripeyellow rust},
   \item \code{Yellow spot} or
   \item \code{White grain}
 }
}

\section{Accepted \code{location_description} Values}{

 Acceptable values for location are
 \itemize{
   \item \code{farm field},
   \item \code{commercial trial},
   \item \code{research trial} or
   \item \code{roadside}.
 }
}

\section{Accepted \code{season} Values}{

 Acceptable values for season are
 \itemize{
   \item \code{Winter 2018},
   \item \code{Summer 2018/19},
   \item \code{Winter 2019} or
   \item \code{Summer 2019/20}.
 }
}

\examples{
\dontrun{
x <- get_fulcrum()
# map only mungbean surveys
x <- filter_fulcrum(fd = x, crop = "mungbean")
}
}
\seealso{
\code{\link{get_fulcrum}} for retrieving and formatting Fulcrum
 data
}
\author{
Sparks, Adam H. \email{adam.sparks@usq.edu.au}
}
