% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fulcrum.R
\name{get_fulcrum}
\alias{get_fulcrum}
\title{Fetches Data from Fulcrum.app and Formats Them for Use in Reporting}
\usage{
get_fulcrum(fulcrum_url = NULL, locations_only = FALSE)
}
\arguments{
\item{fulcrum_url}{A url provided by \url{https://www.fulcrumapp.com/} for
sharing a 'CSV' format fill of Fulcrum data}

\item{locations_only}{Boolean.  If \code{TRUE}, return locations surveyed
along with location and crop metadata but no disease observations.  Defaults
to \code{FALSE}.}
}
\value{
A \code{\link[sf]{sf}} object of Fulcrum survey data projected into
Australia Albers for mapping
}
\description{
Fetch and clean Fulcrum data from USQ/DAFQ in-paddock surveys of crop
disease incidence and oher related notes for reporting.
}
\examples{

\dontrun{
# using a URL set in the .Renviron, get data from Fulcrum
x <- get_fulcrum()
}

\dontrun{
# specify a url to get data from Fulcrum
x <- get_fulcrum(fulcrum_url = "https://web.fulcrumapp.com/shares/########.csv")
}

}
\author{
Sparks, Adam H. \email{adam.sparks@usq.edu.au}
}
