% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fulcrum.R
\name{map_fulcrum}
\alias{map_fulcrum}
\title{Create a Map of Survey Sites}
\usage{
map_fulcrum(fd)
}
\arguments{
\item{fd}{Required.  Fulcrum data to be filtered.  Data must be a result of
\code{\link{get_fulcrum}}.}
}
\value{
A hexbin map of survey locations as a \pkg{\link[ggplot2]{ggplot2}}
object.
}
\description{
Generates a map of surveyed paddocks for all or selected crops or diseases.
Paddocks are not identified to lat/lon, but binned in hexes to avoid accurate
identification being possible.
}
\author{
adam.sparks@usq.edu.au
}
