% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz_label.R
\docType{data}
\name{oz_label}
\alias{oz_label}
\title{Base Map of Queensland and New South Wales for Labeling}
\format{A \code{\link[sf]{sf}} object.}
\usage{
oz_label
}
\description{
Data to provide a base map for reporting, this data provides a layer with
state outlines and is suitable for labeling.
}
\seealso{
\code{\link{oz_outline}}
}
\keyword{datasets}
